requirements.txt
pywin32==305
PyYAML==5.4.1